Bug 2026462 - Follow-up follow-up, change maxBytes param to size_t in nsMsgI18NTruncateUTF8Str(). r=edicharry
Differential Revision: https://phabricator.services.mozilla.com/D290263

--HG--
extra : rebase_source : 764944eb5242d547f59f6e945fa3a9269f619969
extra : amend_source : 3ead01f1ae868ab0069fe37618b379503e33fb46

+++ mailnews/base/src/nsMsgDBFolder.cpp
@@ -5395,7 +5395,7 @@ NS_IMETHODIMP nsMsgDBFolder::GetMsgTextFromStream(
 
   // finally, truncate the string based on aMaxOutputLen
   if (aMsgText.Length() > aMaxOutputLen) {
-    aMsgText = nsMsgI18NTruncateUTF8Str(aMsgText, aMaxOutputLen);
+    aMsgText = nsMsgI18NTruncateUTF8Str(aMsgText, (size_t)aMaxOutputLen);
   }
 
   // Also assign the content type being returned
